Add files via upload
To predict the upward and downward trends of future Volkswagen stock prices using high dimensional stacked Long Short - Term Memory (LSTM) model of recurrent neural network (RNN). 
For this Volkswagen stock price from the beginning of July 1998 to the end of May 2019 is used. Based on the training on price values of 4500 days, the correlations identified by the LSTM of the Volkswagen stock price, the model will try to predict the recent stock prices since 788 business days before the present day.