Add files via upload
To predict the upward and downward trends of future Google stock prices using high dimensional stacked Long Short - Term Memory (LSTM) model of recurrent neural network (RNN). 
For this Google stock price from the beginning of August 2004 to the end of May 2019 is used Based on the training on price values of 3500 days, the correlations identified or captured by the LSTM of the Google stock price, the model will try to predict trend of recent closing stock prices since 216 business days before the present day.