Add files via upload
To predict the upward and downward trends of future Bank of America stock prices using high dimensional stacked Long Short - Term Memory (LSTM) model of recurrent neural network (RNN). 
For this Bank of America stock price from the beginning of January 2000 to the end of May 2019 is used. Based on the training on price values of 4000 days, the correlations identified or captured by the LSTM of the Bank of America stock price, the model will try to predict the trend of recent closing stock prices since 878 business days before the present day. The model is hyper tuned further by reducing the learning rate.