Add files via upload
To predict the upward and downward trends of future Apple stock prices using high dimensional stacked Long Short - Term Memory (LSTM) model of recurrent neural network (RNN). 
For this Apple stock price from the beginning of December 1980 to the end of May 2019 is used. Based on the training on price values of 9500 days, the correlations identified or captured by the LSTM of the Apple stock price, the model will try to predict the trend of recent closing stock prices starting 184 business days ago.