Add files via upload
To predict the upward and downward trends of future Toyota stock prices using high dimensional stacked Long Short - Term Memory (LSTM) model of recurrent neural network (RNN). For this Toyota stock price from the beginning of August 1976 to the end of May 2019 is used. Based on the training on price values of 10600 days, the correlations identified or captured by the LSTM of the Toyota stock price, the model will try to predict the trend of recent closing stock prices starting 186 business days ago. The model is hyper tuned further by reducing the learning rate.