Add files via upload
To predict the upward and downward trends of future Tesla stock prices using high dimensional stacked Long Short - Term Memory (LSTM) model of recurrent neural network (RNN). For this Tesla stock price from the end of June 2010 to the end of May 2019 is used. Based on the training on price values of 2100 days, the correlations identified or captured by the LSTM of the Tesla stock price, the model will try to predict the trend of recent closing stock prices starting 142 business days ago. The model is hyper tuned further by reducing the learning rate.