Add files via upload
To predict the upward and downward trends of future Amazon stock prices using high dimensional stacked Long Short - Term Memory (LSTM) model of recurrent neural network (RNN). For this Amazon stock price from start of May 2010 to the end of May 1997 is used. Based on the training on price values of 5400 days, the correlations identified or captured by the LSTM of the Amazon stock price, the model will try to predict the trend of recent closing stock prices starting 143 business days ago.